refactor(models): relocate providers to shared library
- Moved `getProviderEnvVar` and related logic to shared models directory.
- Updated all references to use the new shared location.
- Simplified imports and removed redundant exports across files.
