+++ apps/gateway/src/api-individual.e2e.ts
@@ -6,10 +6,9 @@ import {
 	logMode,
 	validateLogByRequestId,
 } from "@/chat-helpers.e2e";
-import { getProviderEnvVar } from "@/lib/provider";
 
 import { db, tables, eq } from "@llmgateway/db";
-import { models, providers } from "@llmgateway/models";
+import { models, providers, getProviderEnvVar } from "@llmgateway/models";
 
 import { app } from ".";
 import {
