+++ apps/gateway/src/chat-helpers.e2e.ts
@@ -1,11 +1,10 @@
 import "dotenv/config";
 import { describe, expect, it, type TestOptions } from "vitest";
 
-import { getProviderEnvVar } from "@/lib/provider";
-
 import { db, tables } from "@llmgateway/db";
 import {
 	type ModelDefinition,
+	getProviderEnvVar,
 	models,
 	type ProviderModelMapping,
 	providers,
