+++ apps/gateway/src/test-utils/test-helpers.ts
@@ -1,5 +1,6 @@
 import { redisClient } from "@llmgateway/cache";
 import { db, tables, eq } from "@llmgateway/db";
+import { models, type ProviderModelMapping } from "@llmgateway/models";
 
 // eslint-disable-next-line no-relative-import-paths/no-relative-import-paths
 import { processLogQueue } from "../../../worker/src/worker";
@@ -13,16 +14,29 @@ export async function clearCache() {
 /**
  * Helper function to get test options with retry for CI environment
  */
-export function getTestOptions(): TestOptions {
-	return process.env.CI ? { retry: 3 } : {};
+export function getTestOptions(
+	opts: { completions?: boolean } = {
+		completions: true,
+	},
+): TestOptions {
+	const hasTestOnly = models.some((model) =>
+		model.providers.some(
+			(provider: ProviderModelMapping) => provider.test === "only",
+		),
+	);
+	return process.env.CI || opts?.completions
+		? { retry: 3 }
+		: { skip: hasTestOnly || !!process.env.TEST_MODELS };
 }
 
 /**
  * Helper function to get concurrent test options with retry for CI environment
  * @returns TestOptions with concurrent: true and CI retry configuration
  */
-export function getConcurrentTestOptions(): TestOptions {
-	return { concurrent: true, ...getTestOptions() };
+export function getConcurrentTestOptions(opts?: {
+	completions?: boolean;
+}): TestOptions {
+	return { concurrent: true, ...getTestOptions(opts) };
 }
 
 /**
